refactor(fundraiser): use gatsby filesystem api to create fundraiser pages
