+++ gatsby-node.ts
@@ -2,7 +2,6 @@ import type { GatsbyNode } from 'gatsby'
 
 import resolvers from './gatsby/create-resolvers'
 import schema from './gatsby/customize-schema'
-import { createFundraisersPages } from './gatsby/fundraisers/pages'
 import {
   createFundraiserSchemaCustomization,
   createFundraisersFromMarkdown,
@@ -53,14 +52,6 @@ export const createResolvers: GatsbyNode['createResolvers'] = (args) => {
   resolvers.resolveRegionFields(args)
   resolvers.resolveSubregionFields(args)
   resolvers.resolveTeamMemberFields(args)
-}
-
-/*
-Create Dynamic Pages
-================================================================================
-*/
-export const createPages: GatsbyNode['createPages'] = async (args) => {
-  await createFundraisersPages(args)
 }
 
 /*
